+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import { Geist, Geist_Mono } from 'next/font/google';
-import { METADATA } from '@/configs';
+import { VIEWPORT, METADATA } from '@/configs';
 import { ThemeProvider } from '@/providers/ThemeProvider';
 import { LanguageProvider } from '@/providers/LanguageProvider'
 import './globals.css';
@@ -14,6 +14,7 @@ const geistMono = Geist_Mono({
     subsets: ['latin'],
 });
 
+export const viewport = VIEWPORT;
 export const metadata = METADATA;
 
 const RootLayout = ({children}: {children: React.ReactNode}) => {