+++ src/providers/ThemeProvider.tsx
@@ -20,7 +20,7 @@ const ThemeProvider = ({ children }: { children: React.ReactNode }) => {
     return (
         <NextThemeProvider
             attribute='class'
-            defaultTheme='system'
+            defaultTheme='light'
             enableSystem
         >
             {children}