+++ src/stores/settingStore.ts
@@ -7,10 +7,12 @@ export const useSettingStore = create<SettingStateProps>()(
         (set) => ({
             isMenuSettingOpen: false,
             flagIcon: 'default',
+            flagColor: '#FF0000',
             bombIcon: 'default',
             numberStyle: 'default',
             setIsMenuSettingOpen: (isMenuSettingOpen) => set({ isMenuSettingOpen }),
             setFlagIcon: (flagIcon) => set({ flagIcon }),
+            setFlagColor: (flagColor) => set({ flagColor }),
             setBombIcon: (bombIcon) => set({ bombIcon }),
             setNumberStyle: (numberStyle) => set({ numberStyle }),
         }),
@@ -18,6 +20,7 @@ export const useSettingStore = create<SettingStateProps>()(
             name: 'setting-storage',
             partialize: (state) => ({
                 flagIcon: state.flagIcon,
+                flagColor: state.flagColor,
                 bombIcon: state.bombIcon,
                 numberStyle: state.numberStyle
             }),